# docs/swagger.yaml
schemes:
  - http
swagger: '2.0'
info:
  description: server API for Question Board Application.
  title: Question Board
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
host: localhost:8080
basePath: /api
paths:
...
  /admin/user:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get List of User, for Administrator only.
      tags:
        - User
      summary: Get All User
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/definitions/response.User'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
    post:
      security:
        - ApiKey: []
      description: Route Path for Insert New User, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - User
      summary: Create New User
      parameters:
        - description: send request user code and user name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.User'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
  /admin/user/{id}:
    get:
      security:
        - ApiKey: []
      description: Route Path for Get User Details By ID, for Administrator only.
      tags:
        - User
      summary: Get User By ID
      parameters:
        - type: integer
          description: user id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.MessageData'
              - type: object
                properties:
                  data:
                    $ref: '#/definitions/response.UserDetails'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/user/{id}/delete:
    delete:
      security:
        - ApiKey: []
      description: Route Path for Delete User, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - User
      summary: Delete User
      parameters:
        - type: integer
          description: user id
          name: id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
  /admin/user/{id}/update:
    put:
      security:
        - ApiKey: []
      description: Route Path for Update User, for Administrator only.
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - User
      summary: Update User
      parameters:
        - type: integer
          description: user id
          name: id
          in: path
          required: true
        - description: send request user code and user name
          name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/models.User'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.MessageOnly'
        '417':
          description: Expectation Failed
          schema:
            $ref: '#/definitions/response.Error'
...
definitions:
...
  models.Level:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  models.Major:
    type: object
    properties:
      code:
        type: string
        example: INF
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Informatics
...
  models.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/models.User'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
  models.Tag:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Year
      value:
        type: string
        example: '2019'
  models.TeacherSubject:
    type: object
    properties:
      class:
        type: string
      subject:
        $ref: '#/definitions/models.Subject'
      subject_id:
        type: integer
      teacher:
        $ref: '#/definitions/models.User'
      teacher_id:
        type: integer
...
  models.User:
    type: object
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      level_id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      major_id:
        type: integer
      name:
        type: string
      password:
        type: string
      status:
        type: integer
      subject:
        type: array
        items:
          $ref: '#/definitions/models.Subject'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      teacher_subject:
        type: array
        items:
          $ref: '#/definitions/models.TeacherSubject'
      updatedAt:
        type: string
...
  response.Error:
    type: object
    properties:
      error: {}
      message:
        type: string
  response.MessageData:
    type: object
    properties:
      data: {}
      message:
        type: string
  response.MessageOnly:
    type: object
    properties:
      message:
        type: string
...
  response.Student:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
  response.Subject:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      student:
        type: array
        items:
          $ref: '#/definitions/response.Student'
      teacher_class:
        type: array
        items:
          $ref: '#/definitions/response.TeacherClass'
  response.SubjectTeacher:
    type: object
    properties:
      class:
        type: string
      subject:
        $ref: '#/definitions/response.SubjectWithoutTeacher'
...
  response.SubjectWithoutTeacher:
    type: object
    properties:
      code:
        type: string
      id:
        type: integer
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
  response.Teacher:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
  response.TeacherClass:
    type: object
    properties:
      class:
        type: string
      teacher:
        $ref: '#/definitions/response.Teacher'
  response.User:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      name:
        type: string
      status:
        type: integer
...
  response.UserDetails:
    type: object
    properties:
      email:
        type: string
      id:
        type: integer
      level:
        $ref: '#/definitions/models.Level'
      major:
        $ref: '#/definitions/models.Major'
      name:
        type: string
      status:
        type: integer
      subject:
        type: array
        items:
          $ref: '#/definitions/response.Subject'
      tags:
        type: array
        items:
          $ref: '#/definitions/models.Tag'
      teacher_subject:
        type: array
        items:
          $ref: '#/definitions/response.SubjectTeacher'
securityDefinitions:
  ApiKey:
    type: apiKey
    name: Authorization
    in: header